refactor: Remove duplicated code
